Refactoring the filenames
